requirements.txt
flask
google-cloud-storage
google-generativeai
chromadb
pymupdf
